+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### nova-trame, 1.8.0 (unreleased)
 
 * Adds `nova.trame.get_app` to make it easier to access `ThemedApp` methods from child components (thanks to John Duggan).
+* `InputField` now supports `type="button"` for creating buttons (thanks to John Duggan).
 
 ### nova-trame, 1.7.1